Added capability to build-and-run different Dockerfiles.  Fixed bug exec'ing ubuntu systemd binary.

+++ bin/build-and-run.sh
@@ -2,11 +2,18 @@
 
 set -e
 
-echo "Removing any unit test containers from a previous run."
-podman-compose -f docker-compose.test.yml down
+BUILD="."
+[ -n "$1" ] && BUILD="-f Dockerfile-$1"
 
 echo "Building the application."
-podman build -t lochnerr/systemd .
+podman build -t lochnerr/systemd $BUILD
+if [ "$?" -ne 0 ]; then
+	echo "ERROR: Build failed."
+	exit 1
+fi
+
+echo "Removing any unit test containers from a previous run."
+podman-compose -f docker-compose.test.yml down
 
 echo "Starting the unit test containers."
 podman-compose --podman-run-args='--systemd=always' -f docker-compose.test.yml up
